oolab/src/main/java/agh/ics/oop/model/GrassField.java
package agh.ics.oop.model;
import agh.ics.oop.model.util.RandomPositionGenerator;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

public class GrassField extends AbstractWorldMap {
    private final Map<Vector2d, Grass> grassMap;
    public GrassField(int grassNum){

        this.grassMap = new HashMap<>();

        RandomPositionGenerator randomPositionGenerator = new RandomPositionGenerator(
                (int) Math.sqrt(10 * grassNum),
                (int) Math.sqrt(10 * grassNum), grassNum);
        for (Vector2d grassPosition : randomPositionGenerator) {
            grassMap.put(grassPosition, new Grass(grassPosition));
        }
    }


    @Override
    public WorldElement objectAt(Vector2d position) {
        WorldElement element = super.objectAt(position);
        if (element != null){
            return element;
        }
        return grassMap.get(position);
    }

    @Override
    public String toString() {
        Vector2d bottom = new Vector2d(upperRight.x, upperRight.y);
        Vector2d top = new Vector2d(lowerLeft.x, lowerLeft.y);
        List<WorldElement> elements = getElements();
        for (WorldElement element: elements) {
            bottom = bottom.lowerLeft(element.getPosition());
            top = top.upperRight(element.getPosition());
        }
        return visualizer.draw(bottom, top);
    }

    @Override
    public List<WorldElement> getElements() {
        List<WorldElement> elements = super.getElements();
        elements.addAll(grassMap.values());
        return elements;
    }
}
